Remove all traces of CanCan

+++ app/controllers/api/v1/clouds/drops_controller.rb
@@ -22,7 +22,7 @@ class Api::V1::Clouds::DropsController < Api::V1Controller
 
     @cloud = Cloud.find(params[:cloud_id])
 
-    authorize! :read, @cloud
+    authorize @cloud, :show?
 
     @drops = Drop.only_visable.after_on_topic(@time,@cloud).order_by_topic(@cloud).page(@page).per(10)
 
@@ -42,7 +42,7 @@ class Api::V1::Clouds::DropsController < Api::V1Controller
     @cloud = Cloud.find(params[:cloud_id])
     @query = params[:q] || ""
 
-    authorize! :read, @cloud
+    authorize @cloud, :show?
 
     @drops = Drop.only_visable.fulltext_search(@query, depositable_ids: { any: [@cloud.id.to_s] })
     @drops = Kaminari.paginate_array(@drops).page(@page).per(10)
@@ -60,7 +60,7 @@ class Api::V1::Clouds::DropsController < Api::V1Controller
 
     @cloud = Cloud.find(params[:cloud_id])
     @drop = Drop.find(params[:id])
-    authorize! :destroy, @drop
+    authorize @drop, :destroy?
 
     if @drop.destroy
       render status: 200
